fix library name loading with latest numpy
With the latest numpy, there is a bug where trying to load
"saiga12.so" does not work-- this is simply a programming bug.
However, you can remove the ".so" and the bug isn't triggered,
and it successfully automatically adds the ".so" to it.

darcs-hash:20080821145221-22ea1-0d54f114cb9c544d85775b45e326d5ef1990a3dd.gz